% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accipiter_age.R
\name{accipiter_age}
\alias{accipiter_age}
\title{Estimate the age of juvenile goshawks based on wing length}
\usage{
accipiter_age(
  df = NULL,
  wing = "wing",
  sex = "sex",
  unit = c("mm", "cm"),
  .plot = F,
  decimals = 2,
  .show_model = T
)
}
\arguments{
\item{df}{data frame containing data}

\item{wing}{column name for wing measurements}

\item{sex}{column name giving the sex of individuals or NULL}

\item{unit}{unit of values. By default in mm}

\item{.plot}{logical}

\item{decimals}{decimals}

\item{.show_model}{logical}
}
\description{
Northern Goshawk age is estimated based on a standard growth curve modeled from published data by
Bijlsma, G. 1997: Handleiding veldonderzoek roofvogels.  KNNV Uitgeverij
}
\details{
Estimations are based on a four degree polynomial fit
}
\examples{
## 240 mm wing length
# accipiter_age(data.frame(wing = c(240,240), sex = c("male","female")))
# accipiter_age(data.frame(wing = c(240,240), sex = c(1,0)))

}
\references{
Bijlsma, G. 1997: Handleiding veldonderzoek roofvogels. KNNV Uitgeverij
}
