% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_region.R
\name{assign_region}
\alias{assign_region}
\title{Assign nest to natural region}
\usage{
assign_region(lat = NULL, long = NULL, overlay = NULL, root = "C:/OSGeo4W64")
}
\arguments{
\item{lat}{latitude}

\item{long}{longitude}

\item{overlay}{shape file}

\item{root}{root of QGIS installation. By default attemtps to find it}
}
\description{
Assign nest to natural region
}
