% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdinfo.R
\name{birdinfo}
\alias{birdinfo}
\title{Summarise all data from specific individual}
\usage{
birdinfo(x = NULL, buzzard_db = "RData/buzzard_db.RData",
  out.file = NULL)
}
\arguments{
\item{x}{Ring}

\item{buzzard_db}{path to buzzard_db}

\item{out.file}{valid path to write to text. By default temporary only}
}
\description{
Summarise all data from specific individual
}
