% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.breeders.R
\name{check.breeders}
\alias{check.breeders}
\title{Check for if breeders were not resighted}
\usage{
check.breeders(buzzard_db = "RData/buzzard_db.RData")
}
\arguments{
\item{buzzard_db}{path}
}
\description{
Check for if breeders were not resighted
}
\examples{
## not run
#test <- check.breeders(buzzard_db = "../../01-PhD/00-Raw/RData/buzzard_db.RData")

}
