% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.buzzard_db.R
\name{check.buzzard_db}
\alias{check.buzzard_db}
\title{Checking buzzard_db}
\usage{
check.buzzard_db(input = "RData/buzzard_db.RData",
  output = paste0("Checks/buzzard_db-", Sys.Date(), ".txt"))
}
\arguments{
\item{input}{path to 'buzzard_db.RData'}

\item{output}{destination for output}
}
\value{
text file
}
\description{
Checks entries of the ringing list for potential issues
}
