% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.dupl.R
\name{check.dupl}
\alias{check.dupl}
\title{Identifies and returns duplicated characters}
\usage{
check.dupl(var = NULL, df = buzzard_db[["ring_db"]], not = NULL)
}
\arguments{
\item{var}{variable name}

\item{df}{data frame}

\item{not}{string to ignore}
}
\value{
character vector
}
\description{
Identifies and returns duplicated characters
}
