% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.overdisp.R
\name{check.overdisp}
\alias{check.overdisp}
\title{Checks for overdispersion in model fits}
\source{
https://ase.tufts.edu/gsc/gradresources/guidetomixedmodelsinr/mixed%20model%20guide.html
}
\usage{
check.overdisp(model)
}
\arguments{
\item{model}{fitted model}
}
\value{
p.value for the null hypothesis that data are not overdispersed
}
\description{
Checks for overdispersion in model fits
}
