% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.resights.R
\name{check.resights}
\alias{check.resights}
\title{Check for duplicated resights}
\usage{
check.resights(buzzard_db = "RData/buzzard_db.RData")
}
\arguments{
\item{buzzard_db}{path}
}
\description{
Check for duplicated resights
}
