% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.tag.milvus.R
\name{check.tag.milvus}
\alias{check.tag.milvus}
\title{Identify to which of multiple possible individuals are wingtag belongs}
\usage{
check.tag.milvus(ring_db = NULL, resights = NULL)
}
\arguments{
\item{ring_db}{ringing list}

\item{resights}{data.frame}
}
\description{
Wingtags have been reused after individuals deceased. This function tests which individual is plausible when there are multiple. Additionally tests that wingtag existed at the reported date.
}
