% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckRing.R
\name{CheckRing}
\alias{CheckRing}
\title{Check the format of a RING database}
\usage{
CheckRing(DNS1 = "Ring2017", DNS2 = "Ring2018", Lookup = "Lookup")
}
\arguments{
\item{DNS1}{RING Database used as standard}

\item{DNS2}{RING Database to check}

\item{Lookup}{access database with variable definitions}
}
\description{
Check the format of a RING database
}
