% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.df.R
\name{compare.df}
\alias{compare.df}
\title{Compare DFs}
\usage{
compare.df(df.new = NULL, df.old = NULL, by = NULL)
}
\arguments{
\item{df.new}{data frame}

\item{df.old}{data frame}

\item{by}{grouping column}
}
\description{
Compare DFs
}
