% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord2shape.R
\name{coord2shape}
\alias{coord2shape}
\title{Create shape for a single coordinate}
\usage{
coord2shape(
  lat = NULL,
  long = NULL,
  out = DBChecks::temp.shp(),
  CRS = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  root = "C:/OSGeo4W64",
  EPSG = "EPSG:32632"
)
}
\arguments{
\item{lat}{lat}

\item{long}{long}

\item{out}{By default temp file}

\item{CRS}{CRS for given lat and long}

\item{root}{root of QGIS installation. By default attemtps to find it}

\item{EPSG}{EPSG code}
}
\description{
Create shape for a single coordinate
}
