% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.capture.histories.R
\name{create.ch}
\alias{create.ch}
\title{Create live encounter History for CJS model}
\usage{
create.ch(
  start = 2007,
  end = 2019,
  interval = c("1", "2"),
  ringing_data = ringing_data,
  resights = resights
)
}
\arguments{
\item{start}{first year}

\item{interval}{1 = annual resights; 2 = seasonal resights}

\item{ringing_data}{data frame}

\item{end:}{last year}

\item{resights:}{data frame}
}
\value{
capture history in LDLD...LD format
}
\description{
Create live encounter History for CJS model
}
