% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.capture.histories.R
\name{create.dl}
\alias{create.dl}
\title{Create Dead and live encounter History (LDLD ...) for Burnham mddel}
\usage{
create.dl(
  start = 2007,
  end = 2019,
  ringing_data = ringing_data,
  resights = resights
)
}
\arguments{
\item{start}{first year}

\item{ringing_data}{data frame}

\item{end:}{last year}

\item{resights:}{data frame}
}
\value{
capture history in LDLD...LD format
}
\description{
Create Dead and live encounter History (LDLD ...) for Burnham mddel
}
