% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2shape.R
\name{df2shape}
\alias{df2shape}
\title{Create shape file using data frame as input}
\usage{
df2shape(df = NULL, dsn = NULL, layer = NULL, Long = "E",
  Lat = "N")
}
\arguments{
\item{df}{data frame}

\item{dsn}{output directory}

\item{layer}{layer name}

\item{Long}{Longitude}

\item{Lat}{Latitude}
}
\description{
Create shape file using data frame as input
}
