% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir2nest.R
\name{dir2nest}
\alias{dir2nest}
\title{Dir2Nest}
\usage{
dir2nest(ring_db = NULL, resights = NULL, repro_fledge_db = NULL)
}
\arguments{
\item{ring_db}{ring_db}

\item{resights}{resights}

\item{repro_fledge_db}{repro_fledge_db}
}
\value{
distance to nest in Km
}
\description{
Dir2Nest
}
