% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2centroid.R
\name{dist2centroid}
\alias{dist2centroid}
\title{Distance between Nest and territory centroid}
\usage{
dist2centroid(df = NULL)
}
\arguments{
\item{df}{data frame where all entries belong to the same territory}
}
\value{
vector of distances in km
}
\description{
Distance between Nest and territory centroid
}
