% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy.R
\name{doy}
\alias{doy}
\title{Julian Day for any date}
\usage{
doy(date = Sys.Date())
}
\arguments{
\item{date}{Object of class date or a string of format "YYYY-MM-DD"}
}
\value{
Julian day
}
\description{
Julian Day for any date
}
\examples{
## day of current date
doy()

}
