% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export2ring.R
\name{export2ring}
\alias{export2ring}
\title{Export ringing data for exchange with the Vogelwarte (RING2.0)}
\usage{
export2ring(DNS1 = NULL, DNS2 = NULL, buzzard = NULL, kite = NULL,
  others = NULL, year = NULL, exclude = NULL, RingerNr = NULL)
}
\arguments{
\item{DNS1}{name of odbc database 32-bit with Driver to Microsoft access (*.mdb). Must contain correct VarTypes.}

\item{DNS2}{name of odbc database 32-bit with Driver to Microsoft access (*.mdb).}

\item{buzzard}{path to 'buzzard_db.RData'}

\item{kite}{path to 'Kite ringing list.xlsx'}

\item{others}{path to 'Goshawk and others ringing list.xlsx'}

\item{year}{vector of years to export}

\item{exclude}{ring numbers to exclude from export (i.e. individuals without Vogelwartenring)}

\item{RingerNr}{Ringer Identity issued by Vogelwarte Helgoland}
}
\description{
Prerequisites:
1.) R running as 32 bit version. (Change in RStudio using Tools > Global options)
2.) Existing Microsoft Access Driver (*.mdb) linking to 'Ring2017.mde'
3.) Existing Microsoft Access Driver (*.mdb) linking to 'Ring2018.mde' --> Will be overwritten !!!
4.) Access to buzzard data as list buzzard_db.RData (sources from excel file)
5.) Access to kite data as excel file Kite ringing list.xlsx
6.) Access to goshawk data as ...
}
