% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.select.R
\name{file.select}
\alias{file.select}
\title{Select most recent version of a file}
\usage{
file.select(
  path = NULL,
  pattern = NULL,
  type = NULL,
  decreasing = T,
  not_in_pattern = NULL
)
}
\arguments{
\item{path}{folder to look up files}

\item{pattern}{pattern}

\item{type}{file extension (e.g. ".csv")}

\item{decreasing}{Boolean. If TRUE newest file is selected}

\item{not_in_pattern}{optional. Pattern which is not allowed}
}
\description{
Select most recent version of a file
}
