% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.geno.R
\name{get.geno}
\alias{get.geno}
\title{Get genotype table for focal individual}
\usage{
get.geno(focal = NULL, path = NULL, add.unrelated = T, add.neighbour = T)
}
\arguments{
\item{focal}{Ring number}

\item{path}{to buzzard_db}

\item{add.unrelated}{logical}

\item{add.neighbours}{logical}
}
\description{
For a focal individual (ie known Ring) collect all genotypes for presumed
offspring and if applicable all chicks from the same territories to check
for switches of territory owners.
}
