% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx.add.R
\name{gpx.add}
\alias{gpx.add}
\title{Add nest ID where missing to ringing list}
\usage{
gpx.add(ring = ring_db, repro = repro_fledge_db)
}
\arguments{
\item{ring}{rining list}

\item{repro}{repro & fledge data}
}
\description{
Add nest ID where missing to ringing list
}
