% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx.filt.R
\name{gpx.filt}
\alias{gpx.filt}
\title{Filter GPX formatted waypoint data}
\usage{
gpx.filt(
  input = NULL,
  output = NULL,
  region = NULL,
  add.territory = T,
  buzzard_db = "RData/buzzard_db.RData",
  reduced = F
)
}
\arguments{
\item{input}{input file (*.gpx)}

\item{output}{output file (*.gpx). By default adds. ".filt".}

\item{region}{optional argument defining a geographic grid. data frame with
arguments: "north", "east", "west" & "south"}

\item{add.territory}{optional argument, allows to add territory name of last
breeding attempt to every point. Currently only Buteo buteo.}

\item{buzzard_db}{path to database file}

\item{reduced}{logical}
}
\description{
Removes duplicates from gpx files based on identical coordinates and
elevation. Additional arguments allow to extract gpx points within a
user-defined grid. Requires packages magrittr and readOGR to be installed.
}
\examples{
## Not run
# test <- gpx.filt(input = "../../../Dropbox/Fieldsaison_2019/GPX data/Pooled_GPX_2019-07-11.gpx",                       region = data.frame(north = 52.22, east = 8.52, south = 51.98, west = 8.30))
}
