% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx.get.R
\name{gpx.get}
\alias{gpx.get}
\title{Get corresponding GPX waypoint for a pair of coordinates}
\usage{
gpx.get(gpx.sp = gpx, repro = repro_fledge_db, digits = 5)
}
\arguments{
\item{gpx.sp}{gpx waypoints}

\item{repro}{repro & fledge data}

\item{digits}{desired resolution (i.e. decimal point)}
}
\description{
Get corresponding GPX waypoint for a pair of coordinates
}
