% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.analysis.R
\name{habitat.analysis}
\alias{habitat.analysis}
\title{Habitat analysis}
\usage{
habitat.analysis(features = NULL, overlay = NULL,
  root = "C:/OSGeo4W64")
}
\arguments{
\item{features}{list of ESRI shapes}

\item{overlay}{overlay ESRI shape file}

\item{root}{root of QGIS installation. By default attemtps to find it}
}
\description{
Crops vector layers using overlay and recalculates area
(if polygon) and length (if line) of objects
}
\details{
vectorised if length overlay > 1

1.) Clip to overlay extent (native:clip)
2.) Recalculate geometry of objects
3.) Output data.frame
}
