% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.mens.R
\name{k.means}
\alias{k.means}
\title{K-means}
\usage{
k.means(buzzard_db = "RData/buzzard_db.RData", root = "C:/OSGeo4W64",
  k = 2)
}
\arguments{
\item{buzzard_db}{path to buzzard_db}

\item{root}{root of QGIS installation. By default attemtps to find it}

\item{k}{number of clusters to compute}
}
\description{
K-means
}
