% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.nearest.distr.R
\name{k.nearest.distr}
\alias{k.nearest.distr}
\title{K-nearest neighbour distribution}
\usage{
k.nearest.distr(
  da = data.frame(Id = buzzard_db$repro_fledge_db$Territory, Year =
    buzzard_db$repro_fledge_db$Year, Lat = buzzard_db$repro_fledge_db$N, Lon =
    buzzard_db$repro_fledge_db$E),
  k = 1
)
}
\arguments{
\item{k}{integer or integer vector denoting k.nearest neighbour to compute}

\item{dat}{data frame with columns Id, Year, Lat & Lon}
}
\description{
Computes k-nearest neighbour values within a year
}
