% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milvus_age.R
\name{milvus_age}
\alias{milvus_age}
\title{Estimate the age of juvenile Red Kite based on wing length}
\usage{
milvus_age(x)
}
\arguments{
\item{x}{wing length in cm}
}
\value{
estimated age
}
\description{
Red Kite age is estimated based on a standard growth curve modeled from published data by
Mammen & Stubbe 1995: Alterseinschätzung und Brutbeginn des Rotmilans (Milvus milvus). Vogel und Umwelt Band 8:91:98
}
\references{
Mammen & Stubbe 1995: Alterseinschätzung und Brutbeginn des Rotmilans (Milvus milvus). Vogel und Umwelt Band 8:91:98
}
