% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.features.R
\name{nearest.features}
\alias{nearest.features}
\title{Habitat analysis: Nearest features}
\usage{
nearest.features(
  features = NULL,
  overlay = NULL,
  point = NULL,
  root = "C:/OSGeo4W64"
)
}
\arguments{
\item{features}{list of ESRI shapes}

\item{overlay}{overlay ESRI shape file}

\item{point}{nest coordinate}

\item{root}{root of QGIS installation. By default attemtps to find it}
}
\description{
Calculates the shortest distance between a point and the edges of a feature.
Preprocessing as in habitat.analysis.R
}
\details{
1.) If feature is a polygon convert to lines using SAGA
2.) Create points along lines
3.) find nearest match between points of both layers
}
