% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.neighbour.R
\name{nearest.neighbour}
\alias{nearest.neighbour}
\title{next neighbour}
\usage{
nearest.neighbour(df = NULL, year = NULL, terr.id = NULL, threshold = 1)
}
\arguments{
\item{df}{data frame}

\item{year}{integer}

\item{terr.id}{Terr_ID}

\item{threshold}{threshold im km}
}
\description{
next neighbour
}
