% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nest2polygon.R
\name{Nest2Polygon}
\alias{Nest2Polygon}
\title{Nest2polygon}
\usage{
Nest2Polygon(df = NULL, root = "C:/OSGeo4W64")
}
\arguments{
\item{df}{data frame containing the following variables:
`E` (Long), `N` (Lat) and `Terr_ID` as a grouping variable}

\item{root}{root of QGIS installation. By default attemtps to find it}
}
\value{

}
\description{
Creates polygons from nest coordinates by connecting nests forming the
maximum territory extent by lines. Then, lines are transformed to polygons.
Requires `RQGIS3` and `QGIS`
}
