% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passer2ring.R
\name{passer2ring}
\alias{passer2ring}
\title{Create RING database for Passer domesticus ringing data}
\usage{
passer2ring(DNS1 = NULL, DNS2 = NULL, data = NULL, RingerNr = NULL,
  exclude = NULL)
}
\arguments{
\item{DNS1}{name of odbc database 32-bit with Driver to Microsoft access (*.mdb). Must contain correct VarTypes.}

\item{DNS2}{name of odbc database 32-bit with Driver to Microsoft access (*.mdb).}

\item{data}{data file in .xlxs}

\item{RingerNr}{Ringer Identity issued by Vogelwarte Helgoland}

\item{exclude}{ring numbers to exclude from export (i.e. individuals without Vogelwartenring)}
}
\description{
Create RING database for Passer domesticus ringing data
}
