% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonize_territories.R
\name{polygonise_territories}
\alias{polygonise_territories}
\title{Polygonize territories}
\usage{
polygonise_territories(nests = buzzard_db[["repro_fledge_db"]],
  root = "C:/OSGeo4W64", temp.dir = "Shapes/temp/",
  out = "Shapes/TerrPolygon2.shp")
}
\arguments{
\item{nests}{data frame}

\item{out}{path}
}
\description{
Based on all nest coordinates for a given territory, a polygon area is
created using `Delaunay triangulation`. Geometries are simplified using the
`Dissolve option` in `QGIS3`
}
\details{
Applied to each territory separately to avoid issues caused by overlapping
geometric features. Resulting polygon are then merged within a single
shape file layer. Note, the triangulation and polygonization only works if
there are at least three coordinates available. Hence, if this criterion is
not fullfilled dummy coordinates are added.
}
