% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFormTools.R
\name{ReadFormTools}
\alias{ReadFormTools}
\title{Format FormTools output}
\usage{
ReadFormTools(path = "data", file = "form561", type = ".csv")
}
\arguments{
\item{path}{path}

\item{file}{file pattern}

\item{type}{file type}
}
\description{
Format FormTools output
}
