% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFormTools.xlsx.R
\name{ReadFormTools.xlsx}
\alias{ReadFormTools.xlsx}
\title{Format FormTools output}
\usage{
ReadFormTools.xlsx(path = "data", file = "webform_download",
  type = ".xlsx")
}
\arguments{
\item{path}{path}

\item{file}{file pattern}

\item{type}{file type}
}
\description{
Format FormTools output
}
