% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.duplicated.resights.R
\name{remove.dupl.resight}
\alias{remove.dupl.resight}
\title{remove duplicated entries and discard unknown individuals}
\usage{
remove.dupl.resight(resights = NULL, threshold = 1)
}
\arguments{
\item{resights}{data frame}

\item{threshold}{difference in distance to nest}
}
\value{
data frame
}
\description{
remove duplicated entries and discard unknown individuals
}
