% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resights.report.R
\name{resight.report}
\alias{resight.report}
\title{Report of resightings}
\usage{
resight.report(x = NULL, buzzard_db = "RData/buzzard_db.RData",
  out.file = NULL, digits = 2, lang = c("German", "English"))
}
\arguments{
\item{x}{Ring}

\item{buzzard_db}{path to buzzard_db}

\item{out.file}{valid path to write to text. By default temporary only}

\item{digits}{number of decimals}

\item{lang}{language}
}
\description{
Report of resightings
}
