% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resights.report.all.R
\name{resights.report.all}
\alias{resights.report.all}
\title{Report on all individuals ever reported by a observer}
\usage{
resights.report.all(
  buzzard_db = "RData/buzzard_db.RData",
  digits = 4,
  lang = c("German", "English"),
  brood = F,
  observer = NULL
)
}
\arguments{
\item{buzzard_db}{path to buzzard_db}

\item{digits}{number of decimals}

\item{lang}{language}

\item{brood}{Default False. If True, returns resights of siblings}

\item{observer}{name}
}
\description{
Report on all individuals ever reported by a observer
}
