% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terr.centroid.R
\name{terr.centroid}
\alias{terr.centroid}
\title{Centoid of a territory}
\usage{
terr.centroid(name = NULL, df = buzzard_db[["repro_fledge_db"]],
  weighted = T)
}
\arguments{
\item{name}{Territory name}

\item{df}{data frame}

\item{weighted}{Logical. By default weighted mean of all breeding attempts}
}
\value{
Centroid lon and lat coordinates
}
\description{
Centoid of a territory
}
