% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FledgeFromHatch.R
\name{FledgeFromHatch}
\alias{FledgeFromHatch}
\title{Calculate Fledging date for hatching date}
\usage{
FledgeFromHatch(x = NULL, year = "2019", verbose = T)
}
\arguments{
\item{x}{Julian day}

\item{year}{Year}

\item{verbose}{logical}
}
\value{
Fledging day as June day
}
\description{
Calculate Fledging date for hatching date
}
