% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlaceInfo.R
\name{PlaceInfo}
\alias{PlaceInfo}
\title{Get Location info}
\usage{
PlaceInfo(lat = NULL, long = NULL, root = "C:/OSGeo4W64")
}
\arguments{
\item{lat}{latitude}

\item{long}{longitude}

\item{root}{root of QGIS installation. By default attemtps to find it}
}
\description{
For a given coordinate queries information of the place from country level down to municipality
}
\details{
Uses qgis:joinattributesbylocation to select data of the GADM ESRI shape  at level 3
}
